ch07/agent_framework/agent_with_memory/requirements.txt
agent-framework==1.0.0b260210
pydantic>=2.0.0
python-dotenv>=1.0.1